Remove thousands delimiter from the price
Hello,
you have a great repo with useful and awesome bloomberg quote. I have tried your file to get quotes from bloomberg into the GnuCash. It works great, but I got an error when trying to obtain quotes with higher value than $1,000.00 (e.g XAUUSD:CUR for gold spot). Then GnuCash complained about "System error." I went on and investigated and found out it complains about comma in the price. I went ahead and removed it in the code and now it works. I'm not very well acquainted with Perl, so I did some search on Google and fixed it as best as I could.

Thanks for your work!

Tomas